Make sure this works the first time you run it

+++ setup.py
@@ -22,7 +22,9 @@ def setup_pyquic():
     with temp_cd('pyquic/py_quic'):
         os.system('make')
 
-    shutil.rmtree('quic/py_quic')
+    if os.path.exists('quic/py_quic'):
+        shutil.rmtree('quic/py_quic')
+    
     shutil.copytree('pyquic/py_quic', 'quic/py_quic')
         
 def clean_pyquic():
